feat(controltower): capture config snapshot on subscribe
Due to Cloudformation limitations, we cannot automatically subscribe S3
buckets to our lambda. This makes it harder to trigger an AWS Config
snapshot, since we cannot do so until events are routed appropriately.

This commit works around this problem by installing an EventBridge Rule
which triggers on a `PutBucketNotification` event which targets our SNS
Topic. On trigger, we notify our lambda that to request an AWS Config
snapshot.