fix(controltower): rename SNS topic to match stack name (#17)
Resource names must either be random or at least partially derived from
the stack name, otherwise it is impossible to install multiple copies of
the same stack. Multiple stack copies may be necessary due to forwarding
the same data to multiple datastreams, or as a side effect of a botched
install or uninstall.