fix(collection): generate snapshot rule name (#20)
This commit omits providing a name for the eventbridge rule responsible
for periodically triggering the API snapshot. In the absence of an
explicit name, cloudformation will generate one based on the stack name,
the resource name and a random suffix. This is more robust than using
the stack name, which can exceed the length constraints of Eventbridge.

As part of this change, we had to modify the dependency chain of
affected resources. Previously, we were gating the creation of the rule
on the permission to trigger the lambda. This ensures that the first
trigger would succeed. By removing the dependency, we risk having the
first event trigger prior to being allowed to invoke the lambda. As
such, it may fail. This turns out to be acceptable because we always
use a custom cloudformation resource to invoke the snapshot directly.