feat(collection): allow configuration of lambda name (#21)
This commit allows falling back to a cloudformation generated name for
the lambda function. This can be useful in cases where the stack name
exceeds the length constraints of a Lambda function name.

Given this is a backward incompatible change, we make the default abide
by legacy behavior. Users must reinstall the stack if they need to flip
the option due to the existence of a downstream custom resource tied to
the lambda ARN ("cannot modify service token").