feat: set snapshot action timeout parameter
Allow configuration of the `SNAPSHOT_ACTION_TIMEOUT` environment
variable in our lambda.

Since our cloudformation stack already takes a `LambdaTimeout` parameter
in seconds, I decided to stick to type `Number` for consistency. We pass
in the value to the lambda by appending `s`.

We can bump the overall LambdaTimeout to 2 minutes, but cut the action
timeout to 90s.